+++ src/pip/_internal/metadata/importlib/_dists.py
@@ -3,16 +3,7 @@ import importlib.metadata
 import os
 import pathlib
 import zipfile
-from typing import (
-    Collection,
-    Dict,
-    Iterable,
-    Iterator,
-    Mapping,
-    NamedTuple,
-    Optional,
-    Sequence,
-)
+from typing import Collection, Dict, Iterable, Iterator, Mapping, Optional, Sequence
 
 from pip._vendor.packaging.requirements import Requirement
 from pip._vendor.packaging.utils import NormalizedName, canonicalize_name
@@ -90,12 +81,6 @@ class WheelDistribution(importlib.metadata.Distribution):
             error = f"Error decoding metadata for {wheel}: {e} in {filename} file"
             raise UnsupportedWheel(error)
         return text
-
-
-class RequiresEntry(NamedTuple):
-    requirement: str
-    extra: str
-    marker: str
 
 
 class Distribution(BaseDistribution):
@@ -187,84 +172,17 @@ class Distribution(BaseDistribution):
         # importlib.metadata's EntryPoint structure sasitfies BaseEntryPoint.
         return self._dist.entry_points
 
-    @property
-    def metadata(self) -> email.message.Message:
+    def _metadata_impl(self) -> email.message.Message:
         return self._dist.metadata
 
-    def _iter_requires_txt_entries(self) -> Iterator[RequiresEntry]:
-        """Parse a ``requires.txt`` in an egg-info directory.
-
-        This is an INI-ish format where an egg-info stores dependencies. A
-        section name describes extra other environment markers, while each entry
-        is an arbitrary string (not a key-value pair) representing a dependency
-        as a requirement string (no markers).
-
-        There is a construct in ``importlib.metadata`` called ``Sectioned`` that
-        does mostly the same, but the format is currently considered private.
-        """
-        content = self._dist.read_text("requires.txt")
-        if content is None:
-            return
-        extra = marker = ""  # Section-less entries don't have markers.
-        for line in content.splitlines():
-            line = line.strip()
-            if not line or line.startswith("#"):  # Comment; ignored.
-                continue
-            if line.startswith("[") and line.endswith("]"):  # A section header.
-                extra, _, marker = line.strip("[]").partition(":")
-                continue
-            yield RequiresEntry(requirement=line, extra=extra, marker=marker)
-
-    def _iter_egg_info_extras(self) -> Iterable[str]:
-        """Get extras from the egg-info directory."""
-        known_extras = {""}
-        for entry in self._iter_requires_txt_entries():
-            if entry.extra in known_extras:
-                continue
-            known_extras.add(entry.extra)
-            yield entry.extra
-
     def iter_provided_extras(self) -> Iterable[str]:
-        iterator = (
-            self._dist.metadata.get_all("Provides-Extra")
-            or self._iter_egg_info_extras()
+        return (
+            safe_extra(extra) for extra in self.metadata.get_all("Provides-Extra", [])
         )
-        return (safe_extra(extra) for extra in iterator)
-
-    def _iter_egg_info_dependencies(self) -> Iterable[str]:
-        """Get distribution dependencies from the egg-info directory.
-
-        To ease parsing, this converts a legacy dependency entry into a PEP 508
-        requirement string. Like ``_iter_requires_txt_entries()``, there is code
-        in ``importlib.metadata`` that does mostly the same, but not do exactly
-        what we need.
-
-        Namely, ``importlib.metadata`` does not normalize the extra name before
-        putting it into the requirement string, which causes marker comparison
-        to fail because the dist-info format do normalize. This is consistent in
-        all currently available PEP 517 backends, although not standardized.
-        """
-        for entry in self._iter_requires_txt_entries():
-            if entry.extra and entry.marker:
-                marker = f'({entry.marker}) and extra == "{safe_extra(entry.extra)}"'
-            elif entry.extra:
-                marker = f'extra == "{safe_extra(entry.extra)}"'
-            elif entry.marker:
-                marker = entry.marker
-            else:
-                marker = ""
-            if marker:
-                yield f"{entry.requirement} ; {marker}"
-            else:
-                yield entry.requirement
 
     def iter_dependencies(self, extras: Collection[str] = ()) -> Iterable[Requirement]:
-        req_string_iterator = (
-            self._dist.metadata.get_all("Requires-Dist")
-            or self._iter_egg_info_dependencies()
-        )
         contexts: Sequence[Dict[str, str]] = [{"extra": safe_extra(e)} for e in extras]
-        for req_string in req_string_iterator:
+        for req_string in self.metadata.get_all("Requires-Dist", []):
             req = Requirement(req_string)
             if not req.marker:
                 yield req
